Refactor core logic into modules
