+++ src/index.html
@@ -239,6 +239,10 @@
   <!-- External script loading order is important -->
   <!-- Load list data files first -->
   <script src="all_lists.js"></script>
+  <!-- Load submodules before main script -->
+  <script src="promptUtils.js"></script>
+  <script src="listManager.js"></script>
+  <script src="uiControls.js"></script>
   <!-- Then load main application logic -->
   <script src="script.js"></script>
 </body>
